+++ examples/volumes/main.tf
@@ -9,32 +9,45 @@ provider "oneview" {
 
 # Extracting resources for volume creation
 data "oneview_scope" "scope_obj" {
-  name = "Auto-Scope"
+  name = "test"
 }
 
 data "oneview_storage_pool" "st_pool" {
-  name = "CPG-SSD-AO"
+  name = "CPG-SSD"
 }
 
 data "oneview_storage_volume_template" "st_vt" {
-  name = "RenameDemoStorageTemplate"
+  name = "vol1"
 }
 
 # Creates volume resource
 resource "oneview_volume" "volume" {
   properties {
-    name              = "testvol"
-     /* uncomment the below two line to provide the size and storage pool uri or it will be taken from the volume template*/
-    //storage_pool      = data.oneview_storage_pool.st_pool.uri
-    //size              = 1221225472/
-    provisioning_type = "Thin" 
+    name        = "testvol3"
+    description = "Test Volume"
+    /* uncomment the below two line to provide the size and storage pool uri or it will be taken from the volume template*/
+    storage_pool      = data.oneview_storage_pool.st_pool.uri
+    size              = 1221225472
+    provisioning_type = "Thin"
     is_deduplicated   = false
   }
-  template_name        = "RenameDemoStorageTemplate"
-  template_uri         = data.oneview_storage_volume_template.st_vt.uri
-  name                 = "testvol"
-  description          = "Test Volume"
+  template_name        = "Auto-VolumeTemplate"
   initial_scope_uris   = [data.oneview_scope.scope_obj.uri]
-  provisioned_capacity = "268435456"
+  provisioned_capacity = "1221225472"
 }
+
+#Creates volume resource without providing volume template
+# resource "oneview_volume" "volume" {
+#   properties {
+#     name              = "testvol3"
+#     description = "Test Volume with no template"
+#     storage_pool      = data.oneview_storage_pool.st_pool.uri
+#     size              = 1221225472
+#     provisioning_type = "Thin"
+#     is_deduplicated   = false
+#   }
+#   initial_scope_uris   = [data.oneview_scope.scope_obj.uri]
+#   provisioned_capacity = "1221225472"
+# }
+