+++ oneview/resource_volume.go
@@ -12,6 +12,8 @@
 package oneview
 
 import (
+	"errors"
+
 	"github.com/HewlettPackard/oneview-golang/ov"
 	"github.com/HewlettPackard/oneview-golang/utils"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
@@ -75,6 +77,11 @@ func resourceVolume() *schema.Resource {
 							Optional: true,
 							Computed: true,
 						},
+						"description": {
+							Type:     schema.TypeString,
+							Optional: true,
+							Computed: true,
+						},
 						"storage_pool": {
 							Type:     schema.TypeString,
 							Optional: true,
@@ -103,6 +110,10 @@ func resourceVolume() *schema.Resource {
 							Type:     schema.TypeBool,
 							Optional: true,
 						},
+						"is_shared": {
+							Type:     schema.TypeBool,
+							Optional: true,
+						},
 						"is_compressed": {
 							Type:     schema.TypeBool,
 							Optional: true,
@@ -118,18 +129,14 @@ func resourceVolume() *schema.Resource {
 				Type:     schema.TypeString,
 				Computed: true,
 			},
-			"is_permanent": {
-				Type:     schema.TypeBool,
-				Optional: true,
-				Computed: true,
-			},
+
 			"is_shareable": {
 				Type:     schema.TypeBool,
 				Computed: true,
 			},
 			"name": {
 				Type:     schema.TypeString,
-				Required: true,
+				Optional: true,
 			},
 			"provisioned_capacity": {
 				Type:     schema.TypeString,
@@ -199,32 +206,43 @@ func resourceVolume() *schema.Resource {
 
 func resourceVolumeCreate(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*Config)
+	var (
+		volume         ov.StorageVolume
+		volTemplate    ov.StorageVolumeTemplate
+		storagePoolUri utils.Nstring
+		volumeSize     int
+		volError       error
+	)
 
-	volume := ov.StorageVolume{}
-
-	volTemplate, err1 := config.ovClient.GetStorageVolumeTemplateByName(d.Get("template_name").(string))
-	if err1 != nil {
-		return err1
-	}
 	properties := d.Get("properties").(*schema.Set).List()[0].(map[string]interface{})
+	if val, ok := d.GetOk("template_name"); ok {
 
-	var volumeSize int
-	var storagePoolUri utils.Nstring
+		volTemplate, volError = config.ovClient.GetStorageVolumeTemplateByName(val.(string))
+		if volError != nil {
+			return volError
+		}
+		storagePoolUri = utils.Nstring(volTemplate.StoragePoolUri)
+	} else if properties["storage_pool"] != "" {
+		volTemplate, volError = config.ovClient.GetRootStorageVolumeTemplate(properties["storage_pool"].(string))
+		if volError != nil {
+			return volError
+		}
+		storagePoolUri = utils.Nstring(properties["storage_pool"].(string))
+	} else {
+		return errors.New("neither template nor storage pool is provided")
+	}
 
 	if properties["size"] != 0 {
 		volumeSize = properties["size"].(int)
 	} else {
 		volumeSize = volTemplate.TemplateProperties.Size.Default
-	}
-	if properties["storage_pool"] != "" {
-		storagePoolUri = utils.Nstring(properties["storage_pool"].(string))
-	} else {
-		storagePoolUri = utils.Nstring(volTemplate.StoragePoolUri)
+
 	}
 
 	volumeProperties := ov.Properties{
 		Storagepool:         storagePoolUri,
-		Name:                d.Get("name").(string),
+		Name:                properties["name"].(string),
+		Description:         utils.Nstring(properties["description"].(string)),
 		Size:                volumeSize,
 		ProvisioningType:    properties["provisioning_type"].(string),
 		DataTransferLimit:   properties["data_transfer_limit"].(int),
@@ -233,14 +251,11 @@ func resourceVolumeCreate(d *schema.ResourceData, meta interface{}) error {
 		IsEncrypted:         properties["is_encrypted"].(bool),
 		IsPinned:            properties["is_pinned"].(bool),
 		IsCompressed:        properties["is_compressed"].(bool),
+		IsShareable:         properties["is_shared"].(bool),
 	}
 	volume.Properties = &volumeProperties
 	volume.TemplateURI = volTemplate.URI
 
-	if value, exist := d.GetOk("is_permanent"); exist {
-		val := value.(bool)
-		volume.IsPermanent = &val
-	}
 	if val, ok := d.GetOk("initial_scope_uris"); ok {
 		rawInitialScopeUris := val.(*schema.Set).List()
 		initialScopeUris := make([]utils.Nstring, len(rawInitialScopeUris))
@@ -249,9 +264,8 @@ func resourceVolumeCreate(d *schema.ResourceData, meta interface{}) error {
 		}
 		volume.InitialScopeUris = initialScopeUris
 	}
-
 	err := config.ovClient.CreateStorageVolume(volume)
-	d.SetId(d.Get("name").(string))
+	d.SetId(properties["name"].(string))
 
 	if err != nil {
 		d.SetId("")
@@ -291,7 +305,6 @@ func resourceVolumeRead(d *schema.ResourceData, meta interface{}) error {
 
 	d.Set("device_specific_attributes", &deviceSpecificAttributesTemplates)
 	d.Set("etag", storageVolume.ETAG)
-	d.Set("is_permanent", storageVolume.IsPermanent)
 	d.Set("is_shareable", storageVolume.IsShareable)
 	d.Set("name", storageVolume.Name)
 	d.Set("state", storageVolume.State)
@@ -313,13 +326,12 @@ func resourceVolumeUpdate(d *schema.ResourceData, meta interface{}) error {
 
 	volume := ov.StorageVolume{}
 
-	isPermanent := d.Get("is_permanent").(bool)
 	isShareable := d.Get("is_shareable").(bool)
 
 	volume.Name = d.Get("name").(string)
 	volume.ProvisioningTypeForUpdate = d.Get("provisioning_type").(string)
 	volume.Description = utils.NewNstring(d.Get("description").(string))
-	volume.IsPermanent = &isPermanent
+
 	volume.URI = utils.NewNstring(d.Get("uri").(string))
 	volume.IsShareable = &isShareable